split io related function out
